Refactor progress photo list to improve clarity
