+++ src/Patient/Progress/index.js
@@ -14,7 +14,7 @@ import Key from './Key';
 import PreventOffline from '../../Basics/PreventOffline';
 import Grid from '@material-ui/core/Grid';
 import Tabs from './Tabs';
-import PhotoList from './PhotoList';
+import PhotoList from './PhotoList/';
 
 const useStyles = makeStyles(theme => ({
     container: {
